rename test_sign to sign
